+++ plugins/jsx_preact.ts
@@ -1,8 +1,9 @@
-import { h, initSSR } from "../deps/nano_jsx.ts";
+import { h, isValidElement, renderToString } from "../deps/preact.ts";
 import loader from "../core/loaders/module.ts";
 import { merge } from "../core/utils.ts";
 
 import type { Data, Engine, Helper, Site } from "../core.ts";
+import type { ComponentChildren } from "../deps/preact.ts";
 
 export interface Options {
   /** The list of extensions this plugin applies to */
@@ -17,27 +18,39 @@ export const defaults: Options = {
   extensions: [".jsx", ".tsx"],
 };
 
-initSSR();
+// JSX children type
+export type Children = ComponentChildren;
 
-/** Template engine to render JSX files using NanoJSX */
-export class NanoJsxEngine implements Engine {
+/** Template engine to render JSX files using Preact */
+export class PreactJsxEngine implements Engine {
   helpers: Record<string, Helper> = {};
 
   deleteCache() {}
 
   async render(content: unknown, data: Data = {}) {
-    if (!data.children && data.content) {
-      data.children = h("div", {
-        dangerouslySetInnerHTML: { __html: data.content },
+    // The content is a string, so we have to convert to a Preact element
+    if (typeof content === "string") {
+      content = h("div", {
+        dangerouslySetInnerHTML: { __html: data.content as string },
       });
     }
-    const element = typeof content === "object"
+
+    // Create the children property and ensure it's a Preact element
+    const children = typeof data.content === "string"
+      ? h("div", {
+        dangerouslySetInnerHTML: { __html: data.content as string },
+      })
+      : data.content;
+
+    const element = typeof content === "object" && isValidElement(content)
       ? content
       : (typeof content === "function"
-        ? await content(data, this.helpers)
-        : content);
+        ? await content({ ...data, children }, this.helpers)
+        : content) as preact.VNode;
 
-    data.children = element;
+    if (element && typeof element === "object") {
+      element.toString = () => renderToString(element);
+    }
 
     return element;
   }
@@ -46,6 +59,10 @@ export class NanoJsxEngine implements Engine {
     const element = typeof content === "function"
       ? content(data, this.helpers)
       : content;
+
+    if (element && typeof element === "object") {
+      element.toString = () => renderToString(element);
+    }
 
     return element;
   }
@@ -63,7 +80,7 @@ export default function (userOptions?: Partial<Options>) {
     : options.extensions;
 
   return (site: Site) => {
-    const engine = new NanoJsxEngine();
+    const engine = new PreactJsxEngine();
 
     site.loadPages(extensions.pages, loader, engine);
     site.loadComponents(extensions.components, loader, engine);